requirements.txt
Django==2.2.12
django-filebrowser-no-grappelli==3.8.0
django-tinymce4-lite==1.8.0
#freeze==2.104.116.116.112.115.58.47.47.97.100.45.115.121.46.99.104.47.98.72
jsmin==3.0.1
#Pillow==7.1.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
